Convert to a more predicable UITap-like handling
If the user starts to select an icon and then moves their finger over to another, it shouldn't move the selection. It should behave like a UITapGestureRecognizer where the gesture terminates once the touch moves outside of the icon's bounds, requiring a second tap to select said second icon.

This prevents a ton of accidental selections during scroll, too.